+++ RELEASES.md
@@ -1,3 +1,8 @@
+# 0.2.0
+
+- [New] Added in a way to check for qty changes on the cart screen.
+- [Update] Completely rewrote the front end JavaScript so it's modular and you can import it at your will. Publishes to `resources/js/vendor/shopify`.
+
 # 0.1.2
 
 - [New] Added trait for importing all products